Added support for multiple cljsbuild builds in the same project.clj
This is extremely convenient for doing library development in
ClojureScript. This allows cljsbuild to compile in all 4 optimization
levels at once, for easier testing, or to compile a test suite
alongside the library code.

Each build runs in parallel, with appropriate locking to ensure output
isn't garbled. It seems to work well in both 'once' and 'auto' modes.