fixing :libs filtering bug
I just noticed this when I was going through the code.

The filtering of js-file build dependencies to keep built files out of
the dependancies wasn’t working. This patch fixes it.

To cause the bug just set :libs to your output-dir and look at the
dependency mtimes list. You will see that it contains all the compiled
js files.

This patch fixes that.