Optimizing environment reload
Currently working on clj files in auto mode is difficult as it causes a
complete rebuild from the ground up external dependancies do not need
to be recompiled.

This commit demonstrates a reload procedure that is equivalent to the
current one but is a lot faster because because all compiled
dependancies are not deleted.

The method employed here touches the compiled target files for all cljs
sources (including crossover sources) with an old modification time.
This forces the compiler to rebuild all local class source files.

External dependancies cannot rely on the current project so there is no
need for them to be recompiled.

This patch is still a blunt way to handle this but it is tremendously
more efficient for the most common cases.