fix reloading of clj files in cljs dirs
cljsbuild “auto” produces different results than “once” if a clj file
that is a dependency of a clj based macro file.  This is because
non-macro clj files are not reloaded on change.

The issue behind https://github.com/emezeske/lein-cljsbuild/issues/210
appears to have been resolved as of clojurescript 2202.

I removed the guard that only allows .clj files that contain macros to
be reloaded.  I have tested this and it works pretty darn good :)