Add support for Nashorn & Node.js repls.
Code for all three "embedded" REPLs is now in a single namespace.

This required upgrading the Clojure and ClojureScript versions used for
development and testing, which also required upgrading Midje and
lein-midje.

Test updates are included, though I could not get the tests to run
properly in the `plugin` project - I am not sure how they *could* run
given the current state of the code. Midje seems to bail early instead
of acknowledging an expected error. Tests in `support` pass correctly.