Fix #465: Use full project map for subproject.
Previously, when we would just select a few keys from the project map
for the subproject, we would not include the :root key. The root key
specifies the path to the project directory. It is used to set the
working directory when starting a new subprocess. Not having the root
key in the subproject map would cause problems when using lein-monolith
since the working directory would not be set correctly.